+++ packages/_example/scripts/sequelize-mariadb.ts
