+++ packages/_example/src/forest/datasources/live.ts
@@ -1,7 +1,7 @@
 import LiveDataSource from '@forestadmin/datasource-live/dist/datasource';
 import faker from '@faker-js/faker';
 
-import sequelizeMySql from '../../orms/sequelize-mysql';
+import sequelizeMySql from '../../connections/sequelize-mysql';
 
 // This seed is in this file because the records are loaded in the memory.
 export async function seedLiveDatasource(dataSource: LiveDataSource) {
