+++ packages/_example/src/index.ts
@@ -6,7 +6,7 @@ dotenv.config();
 export default async () => {
   const agent = makeAgent().mountStandalone(Number(process.env.HTTP_PORT_LOCAL));
 
-  await makeAgent().start();
+  await agent.start();
 
   return () => agent.stop();
 };
